+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Adjust get_wan_ip to not require Union
+- Switch configuration loaders to use logger instead of print messages
 
 ### Deprecated